norgesGruppen/training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics==8.1.0",
]
